SQL commands almost done
Set up of the tables in the database should now be compleate, population
with data should be next. added a txt that includes only the commands
that need to be run to set up the database and tables. At the request of
Shaun to remove un-needed information from the set-up.
- Jack